Convert most of our zsh configuration to nix
We still have some parts in a normal config file because of zinit and
the lack of support for it in nix. Maybe we will port it to the native
plugin system. What we should do is move the conditions to nix and split
them up into separate files per system.