Allow to use unstable and stable packages in all our modules
We may start to get this, hopefully. We probably still want
`home-manager` to be unstable to have access to its latest modules. This
means it has to work with unstable nixpkgs, this also seems to mean
`nix-darwin` needs to default to unstable nixpkgs. In our shared
modules we might want to install some packages from stable pkags though.
It seems that we have to pass in `stablePkgs` and `unstablePkgs` which
we define differently for nixos and nix-darwin to have access to both.

This works on macos but has to be validated on linux/nixos.