Actually allow to mix and match stable and unstable
Unsure if the previous commit wasn't correct or something else changed
but now we can actually decide if we want the stable or unstable version
of a package in any of our modules.

This all looks a bit gnarly, if/when we understand all of it we can
hopefully clean this up.