Finally figure out the stable / unstable situation
It turns out that home-manager follows the `nix-darwin` nixpkgs version
on macos (probably because of `useGlobalPkgs = true`) which means we
also have to change it to the stable version to make it not complain
about version mismatches and installing unstable packages.

I think `home-manager` uses different arguments/options depending on how
we use it, on linux (not nixos) it uses `inputs.nixpkgs.follows` which
explains why things were working in that case before. All this is a bit
convoluted but we seem now be able to use stable home-manager but use
some unstable packages (we might want to use all unstable because many
packages are relatively old on stable).

Us putting `home.packages` both in the flake (for `linuxSystem`) and the
"home" configuration in `dominik.nix` made things worse, the idea was to
spot version mismatches when building the flake on different setups but
it created weird dependency(?) issues and "false" errors that we would
not have seen without it.