Finally solve the PATH issue in zsh on macos
It turns out zsh has to be enabled globally when using nix-darwin so
that nix properly adds the profiles to the PATH in the right place, the
same is probably also true on nixos. When we did this manually before it
was appended to the path making it so that programs installed via nix
were not overriding system packages.